total cost fix

+++ src/pages/iv-materials-tfr/Form/MaterialsTransferForm.js
@@ -280,7 +280,7 @@ export default function MaterialsTransferForm({ labels, maxAccess: access, recor
 
       return {
         totalQty: acc?.totalQty + qtyValue,
-        totalCost: (Math.round((parseInt(acc?.totalCost) + totalCostValue) * 100) / 100).toFixed(2),
+        totalCost: (Math.round((parseFloat(acc?.totalCost) + totalCostValue) * 100) / 100).toFixed(2),
         totalWeight: acc?.totalWeight + weightValue
       }
     },
